set Rabi frequency of forbidden transitions to zero
In Cell, coupling strengths specified by Rabi frequency were simulated even
when the transition was dipole forbidden.

This change forcesthe Rabi frequency to zero is the dipole moment (which was
already being calculated) is less than the floating point epsilon of the system.
Also, a warning is issued to the user.